d3d12: Additional fixes for MinGW build
Various fixes for GCC build, including actual bug fixes

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/7722>

+++ subprojects/gst-plugins-bad/sys/d3d12/gstd3d12decoder.cpp
@@ -947,8 +947,8 @@ gst_d3d12_decoder_new_picture_with_size (GstD3D12Decoder * decoder,
     return GST_FLOW_ERROR;
   }
 
-  if (priv->session->coded_width >= width &&
-      priv->session->coded_height >= height) {
+  if (priv->session->coded_width >= (gint) width &&
+      priv->session->coded_height >= (gint) height) {
     return gst_d3d12_decoder_new_picture (decoder, videodec, picture);
   }
